+++ app_ethernet_demo/src/ethernet_conf.h
@@ -1,1 +1,0 @@
-//#define ETHERNET_FILTER_SPECIALIZATION [[distributed(arp_ip_filter, 0)]]