+++ module_ethernet/src/phy_support/ethernet_phy_support.h
@@ -1,8 +1,0 @@
-#ifndef __ethernet_phy_support__
-#define __ethernet_phy_support__
-#include "smi.h"
-typedef out port ? ethernet_reset_port_t;
-
-#include "smsc_phy_support.h"
-
-#endif // __ethernet_phy_support__