+++ module_ethernet_board_support/XR-AVB-LC-BRD/ethernet_board_conf.h
@@ -1,7 +1,0 @@
-#ifndef __ethernet_board_defaults_h__
-#define __ethernet_board_defaults_h__
-
-#define ETHERNET_DEFAULT_TILE tile[1]
-#define ETHERNET_DEFAULT_PHY_ADDRESS 0x0
-
-#endif // __ethernet_board_defaults_h__