+++ tests/test_smi/src/smi_conf.h
@@ -1,3 +1,0 @@
-#define SMI_MDC_BIT 1
-#define SMI_MDIO_BIT 0
-#define SMI_HANDLE_COMBINED_PORTS 1